refactor(sales-api domain): optimize data types, timezone handling and AMQP messages dto - Migrate price fields from BigInteger to primitive long across all entities (Seat, Order, Ticket) and DTOs to map efficiently to PostgreSQL BIGINT(int8). - Replace LocalDateTime with Instant in EventEntity and CreateEventRequestDto to enforce UTC standard and prevent timezone shifts. - Rename sequence generators (events_seq, seats_seq) to match Flyway migration scripts. - Add unique constraint to EventEntity name. - Enrich CreateEventMessageDto with 'placesNumber' and 'seatPrefix' to provide full context for the async seat generation consumer.

+++ sales-api/src/main/java/pl/kacper/sales_api/domain/order/OrderEntity.java
@@ -8,8 +8,6 @@ import pl.kacper.sales_api.domain.BaseEntity;
 import pl.kacper.sales_api.domain.eventTicket.TicketEntity;
 import pl.kacper.sales_api.domain.user.UserEntity;
 
-import java.math.BigInteger;
-import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
@@ -40,10 +38,10 @@ public class OrderEntity extends BaseEntity {
     @Enumerated(EnumType.STRING)
     private OrderStatus orderStatus;
 
-    private BigInteger price;
+    private long price;
 
 
-    public OrderEntity(UserEntity purchaser, String paymentSessionId, OrderStatus orderStatus, BigInteger price) {
+    public OrderEntity(UserEntity purchaser, String paymentSessionId, OrderStatus orderStatus, long price) {
         this.purchaser = purchaser;
         this.paymentSessionId = paymentSessionId;
         this.orderStatus = orderStatus;
